requirements.txt
# HTTP requests
requests>=2.31.0

# JWT handling
PyJWT>=2.8.0

# Environment variables
python-dotenv>=1.0.0

# Type checking
typing-extensions>=4.8.0

# Logging and utilities
python-json-logger>=2.0.7 